# backend/docs/openapi.yml
openapi: 3.0.3
info:
  title: AuraTalk Backend (Phase 1–2)
  version: 0.2.0
  description: |
    Consolidated OpenAPI spec for AuraTalk backend services.
    Phase 2 adds username identity, privacy settings, contacts, and message requests.
    NOTE: This backend stores encrypted message envelopes only (no plaintext messages).
servers:
  - url: http://localhost:8082
    description: gateway (local)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Ok:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string
      required: [ok, error]
...
paths:
...
  /me:
    get:
      tags: [users]
      summary: Get current user (JWT required)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      phone:
                        type: string
                      created_at:
                        type: string
                        format: date-time
                    required: [id, phone, created_at]
                required: [ok, user]
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /me/profile:
    post:
      tags: [users]
      summary: Update profile (JWT required)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                  maxLength: 64
                avatarUrl:
                  type: string
                  format: uri
                  maxLength: 512
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ok"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /devices/keys:
    post:
      tags: [users]
      summary: Upsert device public keys (JWT required)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                deviceId:
                  type: string
                identityKeyPublic:
                  type: string
                registrationId:
                  type: integer
                preKeyBundle:
                  type: object
                  additionalProperties: true
              required: [deviceId, identityKeyPublic, registrationId, preKeyBundle]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ok"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{userId}/devices/keys:
    get:
      tags: [users]
      summary: Get device key bundles for a user (JWT required)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Device keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  devices:
                    type: array
                    items:
                      type: object
                      properties:
                        device_id:
                          type: string
                        identity_key_public:
                          type: string
                        registration_id:
                          type: integer
                        pre_key_bundle:
                          type: object
                          additionalProperties: true
                        updated_at:
                          type: string
                          format: date-time
                required: [ok, devices]
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...